+++ src/numguesser++.cc
@@ -5,10 +5,15 @@
 #include <getopt.h>   // Provides argument parsing
 #include <iostream>   // cerr, cin, cout, ios_base
 #include <string>     // getline(), stoi(), str type
-#include <unistd.h>   // getlogin()
+#include <sys/stat.h>
+#include <sys/types.h>
+#include <unistd.h> // getlogin()
 
+#define br '\n'
+#define log(msg) cout << __FILE__ << ":" << __LINE__ << ": " << msg << br;
+
+namespace fs = std::filesystem;
 using namespace std;
-namespace fs = filesystem;
 typedef string str;
 
 int rand_int, num, diff, num_ans, num_range_min, num_range_max, attempts,
@@ -20,7 +25,10 @@ const struct option longopts[] = {{"read-file", required_argument, 0, 'f'},
                                   {"difficulty", required_argument, 0, 'd'},
                                   {"help", no_argument, 0, 'h'},
                                   {"usage", no_argument, 0, 'h'},
-                                  {"version", no_argument, 0, 'v'}};
+                                  {"version", no_argument, 0, 'v'},
+                                  {"prompt-name", no_argument, 0, 's'},
+                                  {"no-num", no_argument, 0, 'n'},
+                                  {0, 0, 0, 0}};
 
 uint rng(uint min, uint max);
 int diff_choose();
@@ -40,83 +48,80 @@ void interrupt(int signum) {
   exit(signum);
 }
 
-// Print the usage document and exit with status scode
+// Print the usage document and exit with status scode.
 void usage(int scode) {
   cout << R"EOF(Usage: numguesser++ [OPTION...]
     or ng++ [OPTION...]
 
-numguesser++ is a C++ rewrite of numguesser, a random number 
-guessing game originally written in C, with some additional 
-enhancements.
-            
--d, --difficulty 1-3    Chooses the difficulty of the game, 
-                        skipping the difficulty select phase. 
-                        Accepts any value from 1-3 inclusive.
+numguesser++ is a C++ rewrite of numguesser, a random number guessing game
+originally written in C, with some additional enhancements.
 
--s                      Prompts for a custom save name after 
-                        a victory. If this option is not 
-                        passed, the username of the user who 
-                        called the program is used instead.
-                                    
--f, --read-file FILE    Reads from FILE and exits. FILE must 
-                        have the '.scf' extension.
-            
--n                      Prevents the program from displaying 
-                        the correct number after a loss.
-            
--h, --help, --usage     Displays this help document and 
-                        exits.
-            
--v, --version           Displays the program's version string 
-                        and exits.
-                                    
-Mandatory or optional arguments to long options are also 
-mandatory or optional for any corresponding short options.
+-d, --difficulty=1-3    Chooses the difficulty of the game, skipping the 
+                        difficulty select phase. Accepts any value from 1-3 
+                        inclusive.
+
+-s, --prompt-name       Prompts for a custom save name after a victory. If
+                        this option is not passed, the username of the user
+                        who called the program is used instead.
+
+-f, --read-file=FILE    Reads from FILE and exits. FILE must have the '.scf'
+                        extension.
+
+-n, --no-num            Prevents the program from displaying the correct
+                        number after a loss.
+
+-h, --help, --usage     Displays this help document and exits.
+
+-v, --version           Displays the program's version string and exits.
+
+Mandatory or optional arguments to long options are also mandatory or optional
+for any corresponding short options.
 
 Exit Codes:
 0 - Everything worked as expected.
 1 - Invalid difficulty value at selection prompt.
 2 - Unable to seed RNG.
-3 - Failed to open score file or score file does not have 
-    .scf extension.
-4 - Invalid difficulty value passed to -d.
-            
+3 - Failed to open score file or score file does not have .scf extension.
+4 - Invalid argument.
+
 Report bugs to https://github.com/Xatra1/numguesser-plus-plus)EOF"
-       << '\n';
+       << br;
   exit(scode);
 }
 
-// Handle any arguments that are passed to the program and set up SIGINT
-// handling.
+// Parse and handle any arguments that are passed to the program and set up
+// SIGINT handling.
 int main(int argc, char *argv[]) {
-  str ftxt;
-  int index;
+  int longopt_index;
   signal(SIGINT, interrupt);
 
   for (;;) {
-    switch (getopt_long(argc, argv, "d:sf:nhv", longopts, &index)) {
+    switch (getopt_long(argc, argv, "d:sf:nphv", longopts, &longopt_index)) {
 
     case 'd': {
       str arg = optarg;
       size_t pos;
+
       try {
         diff = stoi(arg, &pos);
+
         if (diff < 1 || diff > 3) {
-          cout << "\e[33;1;33mwarning:\e[0m Invalid difficulty value (expected "
-                  "1-3 inclusive).\n";
-          usage(4);
+          log("\e[33;1;33mW:\e[0m Invalid difficulty value (expected 1-3 "
+              "inclusive).");
+          diff = 0;
         }
+
       } catch (invalid_argument const &ex) {
-        cout << "\e[33;1;33mwarning:\e[0m Expected integer for difficulty "
-                "argument.\n";
-        usage(4);
+        log("\e[33;1;33mW:\e[0m Expected integer for difficulty argument.");
+        diff = 0;
+
       } catch (out_of_range const &ex) {
-        cout << "\e[33;1;33mwarning:\e[0m Given integer for difficulty "
-                "argument surpasses integer limit.\n";
-        usage(4);
+        log("\e[33;1;33mW:\e[0m Given integer for difficulty "
+            "argument surpasses integer limit.");
+        diff = 0;
       }
-    }
       continue;
+    }
 
     case 's':
       prompt_for_name = true;
@@ -125,19 +130,25 @@ int main(int argc, char *argv[]) {
     case 'f': {
       fs::path extcheck = optarg;
       ifstream f(optarg);
+
       cout << "Reading score file '" << optarg << "'...\n";
+
       if (f.fail()) {
-        cerr << "\a\e[33;1;31mfatal: Unable to open file " << optarg
-             << " for reading.\e[0m\n";
+        cerr << "\a\e[33;1;31mE:\e[0m Unable to open file " << optarg
+             << " for reading.\n";
         return 3;
+
       } else if (extcheck.extension() != ".scf") {
-        cerr << "\a\e[33;1;31mfatal: Score file has file extension "
-             << extcheck.extension() << " (expected .scf)\e[0m\n";
+        cerr << "\a\e[33;1;31mE:\e[0m Score file has file extension "
+             << extcheck.extension() << " (expected .scf)\n";
         return 3;
       }
+
+      str ftxt;
       while (getline(f, ftxt))
-        cout << ftxt << '\n';
+        cout << ftxt << br;
       f.close();
+
       return 0;
     }
 
@@ -150,17 +161,18 @@ int main(int argc, char *argv[]) {
       break;
 
     case 'v':
-      cout << ver << '\n';
+      cout << ver << br;
       return 0;
       break;
 
     default:
-      usage(1);
+      usage(4);
       break;
 
     case -1:
       break;
     }
+
     break;
   }
   rng_seed();
@@ -168,16 +180,19 @@ int main(int argc, char *argv[]) {
 }
 
 // Read a single byte from /dev/urandom, convert that to an integer, and seed
-// RNG.
+// RNG with it.
 void rng_seed() {
   ifstream f("/dev/urandom");
+
   if (f.fail()) {
-    cerr << "\a\e[33;1;31mfatal: Unable to open /dev/urandom\e[0m\n";
+    log("\a\e[33;1;31mE:\e[0m Unable to open /dev/urandom");
     exit(2);
   }
+
   f.get(rand_byte);
   rand_int = int(rand_byte);
   f.close();
+
   srand(rand_int);
 }
 
@@ -188,34 +203,44 @@ int diff_choose() {
             "attempts)\n(3) - Hard (1 attempt)\nMake a selection: ";
     cin >> diff;
   }
+
   switch (diff) {
+
   case 1:
     attempts = 10;
     diff_str = "Easy";
     break;
+
   case 2:
     attempts = 5;
     diff_str = "Normal";
     break;
+
   case 3:
     attempts = 1;
     diff_str = "Hard";
     break;
+
+  case 'n':
+    exit(0);
+
   default:
-    cerr << "\a\n\e[33;1;31mfatal: Invalid difficulty value.\e[0m\n";
+    log("\a\n\e[33;1;31mE:\e[0m Invalid difficulty value.");
     exit(1);
-    break;
   }
+
   cout << "\nYou selected \e[33;1;37m" << diff_str
        << ".\e[0m You will have \e[33;1;37m" << attempts
        << " attempt(s)\e[0m to get the number correct.\n Is this correct? "
           "(y/n): ";
   cin >> ans;
+
   if (ans == 'n')
     while (ans != 'y') {
       diff = 0;
       diff_choose();
     }
+
   return rng_set();
 }
 
@@ -235,13 +260,16 @@ void game() {
     cout << "\nNumber is between " << num_range_min << " and " << num_range_max
          << ". (Attempts left: " << attempts << ")\n Answer: ";
     cin >> num_ans;
+
     if (num_ans == num) {
       attempts--;
       attempts_taken++;
       cout << "\a\e[33;1;37mCorrect!\nAttempts taken: " << attempts_taken
            << "\e[0m\n";
+
       file_ask();
     } else {
+
       while (num_ans != num && attempts > 0) {
         attempts--;
         attempts_taken++;
@@ -249,9 +277,11 @@ void game() {
       }
     }
   } else {
+
     if (show_num)
       cout << "\a\e[33;1;37m\nYou ran out of attempts!\nThe correct number was "
            << num << ".\e[0m\n";
+
     else
       cout << "\a\e[33;1;37m\nYou ran out of attempts!\n\e[0m";
   }
@@ -263,23 +293,30 @@ void file_ask() {
   cout << "Do you want to write information about your run to a score file? "
           "(y/n): ";
   cin >> ans;
+
   if (ans == 'y') {
     ofstream f("scores.scf", ios_base::app); // Append, don't overwrite.
+
     if (f.fail()) {
-      cerr << "\a\e[33;1;31mfatal: Unable to write to score file. Do you have "
-              "write permissions here?\e[0m\n";
+      log("\a\e[33;1;31mE:\e[0m Unable to write to score file. Do you have "
+          "write permissions here?");
       exit(2);
     }
+
     if (prompt_for_name) {
       cout << "Please enter the name you would like to use for your save: ";
       cin >> fname;
     } else
       fname = getlogin();
+
     f << "\t" << fname << "\t\t\t\tAttempts taken: " << attempts_taken
       << "\t\t\t\tAttempts left: " << attempts
-      << "\t\t\t\tDifficulty: " << diff_str << '\n';
+      << "\t\t\t\tDifficulty: " << diff_str << br;
+
     f.close();
+
     cout << "\a\e[33;1;37mInformation written to 'scores.scf'\e[0m\n";
   }
+
   exit(0);
 }